Backend/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
cachelib==0.13.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-Mail==0.10.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
msgspec==0.19.0
PyMySQL==1.1.2
python-dotenv==1.2.1
redis==6.4.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
